Revert 1ba1eb52d6bcfcc9eb8b8a578c4d4427039633fe, added connect method
The connect method breaks the entire assumptions of the class. With
the socket created isValid will return true meaning read/write can
be called. However the socket is not connected.

This is a problem as it's become unclear as to state of the socket.
For now we revert the incrimiating change till we can get the correct
way of doing it.