camera: Overhaul of the camera api & dc1394 / error handling
All functions which previously returned a bool in the camera API now throw
exceptions instead. This makes the code much more robust in the long run as
programmers are slack at error handling and not checking the results from
many of the camera functions will cause crashes elsewhere in their code.
By throwing an exception they at least have to work out why it was thrown.
It also allows a large chunk of errors to all be caught in the same
block of code.

The camera api has had format upgrades so that it supports both v4l and
dc1394. Likewise many of the controls that the api provided have
had minor changes or additions to them. Cameras are also now
required to have a startup method which actually starts frame capturing.

All other camera types have been updated to support the api.

DC1394 has got a major overhaul. The class has now been cleaned up
removing excess exit() functions and to simplify the actual internals
of the class. The camera class also now requires a guid to instanciate
a camera. The DC1394CameraFactory (new class) provides a means of finding
cameras/guids that exists.

There is still a few small things to do to finish the cleanup. The factory
setup is still very early on and whilst it works, we need a way to make a
camera ID unique per camera - thereby allowing easy addressing of the same
camera. We also need a way of getting what formats a camera supports.