tracking/examples: Update ARToolkitPlus tracker to a more working state
We update the constructor of the tracker to support specifing a few more values,
markerwidth, image threshold value. We also define
WCL_ARTOOLKITPLUSTRACKER_MARKER_DETECTION_COUNT, something that can be used to
define how many markers ARP tries to find in a single frame. Sadly this is a
compile time option only.

Update some comments, use dht Robose Pose estimator by default.
Fix up a comment that indicates why a sigsegv occurs upon artoolkitplus tracker
deletion - and how we work around it.

Support the MONO8 camera format. We also no longer store ARP cameras in the
class as ARP manages the memory. At present we specify a hard coded distoriton
matrix as this works. There's work under way to fix this.

Ask the tracker for all markers and also update the pose matrix for the best
marker. We also fix up the row/column major modes when copying marker
information.

Add the ability to set the threshold value for an image, or use
autothresholding.

Update tostring to print out all markers found and the matrix of the best
marker.

Add a way to get the projection matrix and model view matrix for the best
detected marker.

We also update the example artoolkit program to show the marker. It's not
complete yet and should be fixed in an upcoming commit