camera: Provide a getFrame(ImageFormat) method
This method allows the caller to request a particular image format and software
conversion is done from the existing camera format to the requested format. The
image size is maintained. At present only a few conversions are implemented but
they can be easily expanded and an assert is thrown (NULL returned on NDEBUG) if
a requested conversion is tried but is not implemented