tracking: Support near, far planes, don't use uninited camera, allow unsupported video, use default distortion if none exists
Previously the near and far planes were hard coded in ARToolkitPlusTracker. We
now expose these planes so they can be set in the constructor.

We also check that update is not called before setCamera as it causes crashes or
unexpected results.

We now support all video modes via the getFrame(Camera::ImageFormat) method of
Camera. If ARToolkit doesn't natively support a mode, we convert to a format
taht ARToolkit does support (RGB8).

Instead of hardcoding Camera Distortion, we now support getting the distortion
values from the Camera class. If there is no distortion values present, we use
some that are known to work, but may be suboptimal.