camera: Update printDetails, fix 1394 for Configuration, UVC fixes, example fixes
This commit fixes DC1394Camera to work with the new Camera::Configuration setup.
It also unifies printDetails of all Camera classes. PrintDetails now is done in
Camera so that it has a consistent output. It has also been modified to support
a bool indicating if you just want quick details about the camera or full
details about all supported modeds, features etc.

We also update the factories to call printDetails correctly and we update
UVCCamera factory to indicate when cameras can't be found.. partial fix only.

UVCCamera also has the READ/WRITE fileio test removed as we only support mmap
and a fix in place for when a camera doesn't report any formats as valid.

Finally we update the example to allow passing in of a device and printing out
all camera's that exist then the details of the camera that is being used.