Revert "graycode: When building the mask require all image sets to be valid rather than one"
This reverts commit f04c2c07e45079f08ca2cae7845bcc593d216721.

The logic in this is wrong. When building the mask we work with image pairs. The
image pairs must always be black/white combos as we project the image then the
invert. Between pairs however it doesn't need to change. Ie the sequence

black black black (ie all low value pixels)

will invalidate the mask in the current state. Hence we must OR the mask bits
not AND.