Updated system-prep.sh from lucid to natty.
Testing on some boxes seems to indicate that the previous line was
not working for 11.04. However, I have not verified this, and my
tests still work.
The suggestion to change the method for addition of the repository
to add-apt-repository may work in the future; however, this solution
is NOT viable for previous versions of ubuntu (because add-apt-repository
is not installed by default).

The method used (append the deb to /etc/apt/sources.list, install,
then remove the sources.list line) will work for versions of ubuntu
previous to 11.10 for certain.