BoundingBox::intersect functions for Line and Ray.
wcl::BoundingBox can now do boolean intersections with wcl::Line and
wcl::Ray.
We create a LineSegment from the {Line, Ray} using DBL_MAX*direction in
both directions to obtain the two end-points for the segments. This
allows the use of wcl::BoundingBox intersect(wcl::LineSegment), but is
possibly inefficient.