Give OBJGeometry a destructor and move code from OBJParser into it.
OK. So for some reason Ben decided long ago that OBJParser should be
responsible for freeing up the resources used by an OBJGeometry object. Here's
the comment from the code that justified this:

    // Free up the OBJGeometry and all components
    // We don't add this to the structs as we are trying to keep
    // them extremely clean

Ben, if you're reading this then I'm sorry, but you're wrong. Totally wrong.

Why should a file parser be responsible for freeing up the actual Geometry
data? How messy is the struct really going to get by having some implementation
in an external cpp file?

It means you can't give OBJGeometry a working copy constructor. It means you
can't programatically generate an OBJGeometry object. It means you can't ever
declare an OBJParser on the stack in a method and load an OBJ into a member
variable. You have to always put your parser on the heap, and you have to keep
it around for the life of your program.

If you love something, set it free. It's time Parser stopped holding
OBJGeometry back, and stopped trying to be responsible for OBJGeometry's life
choices. From now on, OBJGeometry can live it's own life. Independently of
OBJParser.

It's what's best for both of them. They'll always remain friends.